Create separate classes for groups and dms (browser)
