+++ sdks/browser-sdk/src/Conversation.ts
@@ -1,11 +1,6 @@
 import type { ContentTypeId } from "@xmtp/content-type-primitives";
 import { ContentTypeText } from "@xmtp/content-type-text";
-import type {
-  ConsentState,
-  MetadataField,
-  PermissionPolicy,
-  PermissionUpdateType,
-} from "@xmtp/wasm-bindings";
+import type { ConsentState } from "@xmtp/wasm-bindings";
 import { v4 } from "uuid";
 import { AsyncStream, type StreamCallback } from "@/AsyncStream";
 import type { Client } from "@/Client";
@@ -22,12 +17,6 @@ export class Conversation {
 
   #id: string;
 
-  #name?: SafeConversation["name"];
-
-  #imageUrl?: SafeConversation["imageUrl"];
-
-  #description?: SafeConversation["description"];
-
   #isActive?: SafeConversation["isActive"];
 
   #addedByInboxId?: SafeConversation["addedByInboxId"];
@@ -36,10 +25,6 @@ export class Conversation {
 
   #createdAtNs?: SafeConversation["createdAtNs"];
 
-  #admins: SafeConversation["admins"] = [];
-
-  #superAdmins: SafeConversation["superAdmins"] = [];
-
   constructor(client: Client, id: string, data?: SafeConversation) {
     this.#client = client;
     this.#id = id;
@@ -47,55 +32,14 @@ export class Conversation {
   }
 
   #syncData(data?: SafeConversation) {
-    this.#name = data?.name ?? "";
-    this.#imageUrl = data?.imageUrl ?? "";
-    this.#description = data?.description ?? "";
     this.#isActive = data?.isActive ?? undefined;
     this.#addedByInboxId = data?.addedByInboxId ?? "";
     this.#metadata = data?.metadata ?? undefined;
     this.#createdAtNs = data?.createdAtNs ?? undefined;
-    this.#admins = data?.admins ?? [];
-    this.#superAdmins = data?.superAdmins ?? [];
   }
 
   get id() {
     return this.#id;
-  }
-
-  get name() {
-    return this.#name;
-  }
-
-  async updateName(name: string) {
-    await this.#client.sendMessage("updateGroupName", {
-      id: this.#id,
-      name,
-    });
-    this.#name = name;
-  }
-
-  get imageUrl() {
-    return this.#imageUrl;
-  }
-
-  async updateImageUrl(imageUrl: string) {
-    await this.#client.sendMessage("updateGroupImageUrlSquare", {
-      id: this.#id,
-      imageUrl,
-    });
-    this.#imageUrl = imageUrl;
-  }
-
-  get description() {
-    return this.#description;
-  }
-
-  async updateDescription(description: string) {
-    await this.#client.sendMessage("updateGroupDescription", {
-      id: this.#id,
-      description,
-    });
-    this.#description = description;
   }
 
   get isActive() {
@@ -124,120 +68,12 @@ export class Conversation {
     });
   }
 
-  get admins() {
-    return this.#admins;
-  }
-
-  get superAdmins() {
-    return this.#superAdmins;
-  }
-
-  async listAdmins() {
-    const admins = await this.#client.sendMessage("getGroupAdmins", {
-      id: this.#id,
-    });
-    this.#admins = admins;
-    return admins;
-  }
-
-  async listSuperAdmins() {
-    const superAdmins = await this.#client.sendMessage("getGroupSuperAdmins", {
-      id: this.#id,
-    });
-    this.#superAdmins = superAdmins;
-    return superAdmins;
-  }
-
-  async permissions() {
-    return this.#client.sendMessage("getGroupPermissions", {
-      id: this.#id,
-    });
-  }
-
-  async updatePermission(
-    permissionType: PermissionUpdateType,
-    policy: PermissionPolicy,
-    metadataField?: MetadataField,
-  ) {
-    return this.#client.sendMessage("updateGroupPermissionPolicy", {
-      id: this.#id,
-      permissionType,
-      policy,
-      metadataField,
-    });
-  }
-
-  async isAdmin(inboxId: string) {
-    const admins = await this.listAdmins();
-    return admins.includes(inboxId);
-  }
-
-  async isSuperAdmin(inboxId: string) {
-    const superAdmins = await this.listSuperAdmins();
-    return superAdmins.includes(inboxId);
-  }
-
   async sync() {
     const data = await this.#client.sendMessage("syncGroup", {
       id: this.#id,
     });
     this.#syncData(data);
-  }
-
-  async addMembers(accountAddresses: string[]) {
-    return this.#client.sendMessage("addGroupMembers", {
-      id: this.#id,
-      accountAddresses,
-    });
-  }
-
-  async addMembersByInboxId(inboxIds: string[]) {
-    return this.#client.sendMessage("addGroupMembersByInboxId", {
-      id: this.#id,
-      inboxIds,
-    });
-  }
-
-  async removeMembers(accountAddresses: string[]) {
-    return this.#client.sendMessage("removeGroupMembers", {
-      id: this.#id,
-      accountAddresses,
-    });
-  }
-
-  async removeMembersByInboxId(inboxIds: string[]) {
-    return this.#client.sendMessage("removeGroupMembersByInboxId", {
-      id: this.#id,
-      inboxIds,
-    });
-  }
-
-  async addAdmin(inboxId: string) {
-    return this.#client.sendMessage("addGroupAdmin", {
-      id: this.#id,
-      inboxId,
-    });
-  }
-
-  async removeAdmin(inboxId: string) {
-    return this.#client.sendMessage("removeGroupAdmin", {
-      id: this.#id,
-      inboxId,
-    });
-  }
-
-  async addSuperAdmin(inboxId: string) {
-    return this.#client.sendMessage("addGroupSuperAdmin", {
-      id: this.#id,
-      inboxId,
-    });
-  }
-
-  async removeSuperAdmin(inboxId: string) {
-    return this.#client.sendMessage("removeGroupSuperAdmin", {
-      id: this.#id,
-      inboxId,
-    });
+    return data;
   }
 
   async publishMessages() {
@@ -306,12 +142,6 @@ export class Conversation {
     });
   }
 
-  async dmPeerInboxId() {
-    return this.#client.sendMessage("getDmPeerInboxId", {
-      id: this.#id,
-    });
-  }
-
   async messageDisappearingSettings() {
     return this.#client.sendMessage("getGroupMessageDisappearingSettings", {
       id: this.#id,
@@ -344,11 +174,17 @@ export class Conversation {
     const endStream = this.#client.handleStreamMessage<SafeMessage>(
       streamId,
       (error, value) => {
+        if (error) {
+          void asyncStream.callback(error, undefined);
+          void callback?.(error, undefined);
+          return;
+        }
+
         const decodedMessage = value
           ? new DecodedMessage(this.#client, value)
           : undefined;
-        void asyncStream.callback(error, decodedMessage);
-        void callback?.(error, decodedMessage);
+        void asyncStream.callback(null, decodedMessage);
+        void callback?.(null, decodedMessage);
       },
     );
     await this.#client.sendMessage("streamGroupMessages", {
