+++ sdks/browser-sdk/src/Conversations.ts
@@ -6,8 +6,9 @@ import {
 import { v4 } from "uuid";
 import { AsyncStream, type StreamCallback } from "@/AsyncStream";
 import type { Client } from "@/Client";
-import { Conversation } from "@/Conversation";
 import { DecodedMessage } from "@/DecodedMessage";
+import { Dm } from "@/Dm";
+import { Group } from "@/Group";
 import type {
   SafeConsent,
   SafeConversation,
@@ -38,7 +39,12 @@ export class Conversations {
     const data = await this.#client.sendMessage("getConversationById", {
       id,
     });
-    return data ? new Conversation(this.#client, id, data) : undefined;
+    if (data) {
+      return data.metadata.conversationType === "group"
+        ? new Group(this.#client, data.id, data)
+        : new Dm(this.#client, data.id, data);
+    }
+    return undefined;
   }
 
   async getMessageById(id: string) {
@@ -52,7 +58,7 @@ export class Conversations {
     const data = await this.#client.sendMessage("getDmByInboxId", {
       inboxId,
     });
-    return data ? new Conversation(this.#client, data.id, data) : undefined;
+    return data ? new Dm(this.#client, data.id, data) : undefined;
   }
 
   async list(options?: SafeListConversationsOptions) {
@@ -60,9 +66,10 @@ export class Conversations {
       options,
     });
 
-    return conversations.map(
-      (conversation) =>
-        new Conversation(this.#client, conversation.id, conversation),
+    return conversations.map((conversation) =>
+      conversation.metadata.conversationType === "group"
+        ? new Group(this.#client, conversation.id, conversation)
+        : new Dm(this.#client, conversation.id, conversation),
     );
   }
 
@@ -74,8 +81,7 @@ export class Conversations {
     });
 
     return conversations.map(
-      (conversation) =>
-        new Conversation(this.#client, conversation.id, conversation),
+      (conversation) => new Group(this.#client, conversation.id, conversation),
     );
   }
 
@@ -87,8 +93,7 @@ export class Conversations {
     });
 
     return conversations.map(
-      (conversation) =>
-        new Conversation(this.#client, conversation.id, conversation),
+      (conversation) => new Dm(this.#client, conversation.id, conversation),
     );
   }
 
@@ -98,7 +103,7 @@ export class Conversations {
       options,
     });
 
-    return new Conversation(this.#client, conversation.id, conversation);
+    return new Group(this.#client, conversation.id, conversation);
   }
 
   async newGroupByInboxIds(
@@ -110,7 +115,7 @@ export class Conversations {
       options,
     });
 
-    return new Conversation(this.#client, conversation.id, conversation);
+    return new Group(this.#client, conversation.id, conversation);
   }
 
   async newDm(accountAddress: string, options?: SafeCreateDmOptions) {
@@ -119,7 +124,7 @@ export class Conversations {
       options,
     });
 
-    return new Conversation(this.#client, conversation.id, conversation);
+    return new Dm(this.#client, conversation.id, conversation);
   }
 
   async newDmByInboxId(inboxId: string, options?: SafeCreateDmOptions) {
@@ -128,27 +133,39 @@ export class Conversations {
       options,
     });
 
-    return new Conversation(this.#client, conversation.id, conversation);
+    return new Dm(this.#client, conversation.id, conversation);
   }
 
   async getHmacKeys() {
     return this.#client.sendMessage("getHmacKeys", undefined);
   }
 
-  async stream(
-    callback?: StreamCallback<Conversation>,
+  async stream<T extends Group | Dm = Group | Dm>(
+    callback?: StreamCallback<T>,
     conversationType?: ConversationType,
   ) {
     const streamId = v4();
-    const asyncStream = new AsyncStream<Conversation>();
+    const asyncStream = new AsyncStream<T>();
     const endStream = this.#client.handleStreamMessage<SafeConversation>(
       streamId,
       (error, value) => {
-        const conversation = value
-          ? new Conversation(this.#client, value.id, value)
-          : undefined;
-        void asyncStream.callback(error, conversation);
-        void callback?.(error, conversation);
+        if (error) {
+          void asyncStream.callback(error, undefined);
+          void callback?.(error, undefined);
+          return;
+        }
+
+        let streamValue: T | undefined = undefined;
+
+        if (value) {
+          streamValue =
+            value.metadata.conversationType === "group"
+              ? (new Group(this.#client, value.id, value) as T)
+              : (new Dm(this.#client, value.id, value) as T);
+        }
+
+        void asyncStream.callback(null, streamValue);
+        void callback?.(null, streamValue);
       },
     );
     await this.#client.sendMessage("streamAllGroups", {
@@ -164,12 +181,12 @@ export class Conversations {
     return asyncStream;
   }
 
-  async streamGroups(callback?: StreamCallback<Conversation>) {
-    return this.stream(callback, ConversationType.Group);
+  async streamGroups(callback?: StreamCallback<Group>) {
+    return this.stream<Group>(callback, ConversationType.Group);
   }
 
-  async streamDms(callback?: StreamCallback<Conversation>) {
-    return this.stream(callback, ConversationType.Dm);
+  async streamDms(callback?: StreamCallback<Dm>) {
+    return this.stream<Dm>(callback, ConversationType.Dm);
   }
 
   async streamAllMessages(
@@ -181,11 +198,17 @@ export class Conversations {
     const endStream = this.#client.handleStreamMessage<SafeMessage>(
       streamId,
       (error, value) => {
+        if (error) {
+          void asyncStream.callback(error, undefined);
+          void callback?.(error, undefined);
+          return;
+        }
+
         const decodedMessage = value
           ? new DecodedMessage(this.#client, value)
           : undefined;
-        void asyncStream.callback(error, decodedMessage);
-        void callback?.(error, decodedMessage);
+        void asyncStream.callback(null, decodedMessage);
+        void callback?.(null, decodedMessage);
       },
     );
     await this.#client.sendMessage("streamAllMessages", {
